+++ src/vs/base/common/dataTransfer.ts
