+++ src/vs/workbench/browser/dnd.ts
@@ -3,45 +3,37 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { localize } from 'vs/nls';
-import { IDialogService } from 'vs/platform/dialogs/common/dialogs';
-import { VSBuffer } from 'vs/base/common/buffer';
-import Severity from 'vs/base/common/severity';
-import { IWorkspaceFolderCreationData, IWorkspacesService } from 'vs/platform/workspaces/common/workspaces';
-import { basename, isEqual } from 'vs/base/common/resources';
-import { ByteSize, IFileService } from 'vs/platform/files/common/files';
-import { IWindowOpenable } from 'vs/platform/window/common/window';
-import { URI } from 'vs/base/common/uri';
-import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
-import { FileAccess, Schemas } from 'vs/base/common/network';
-import { IBaseTextResourceEditorInput } from 'vs/platform/editor/common/editor';
 import { DataTransfers, IDragAndDropData } from 'vs/base/browser/dnd';
-import { DragMouseEvent } from 'vs/base/browser/mouseEvent';
-import { Mimes } from 'vs/base/common/mime';
-import { isWeb, isWindows } from 'vs/base/common/platform';
-import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
-import { IEditorIdentifier, GroupIdentifier, isEditorIdentifier, EditorResourceAccessor } from 'vs/workbench/common/editor';
-import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
-import { Disposable, IDisposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { addDisposableListener, DragAndDropObserver, EventType } from 'vs/base/browser/dom';
-import { IEditorGroup } from 'vs/workbench/services/editor/common/editorGroupsService';
-import { IWorkspaceEditingService } from 'vs/workbench/services/workspaces/common/workspaceEditing';
-import { IHostService } from 'vs/workbench/services/host/browser/host';
-import { Emitter } from 'vs/base/common/event';
-import { coalesce } from 'vs/base/common/arrays';
-import { parse, stringify } from 'vs/base/common/marshalling';
-import { ILabelService } from 'vs/platform/label/common/label';
-import { hasWorkspaceFileExtension, isTemporaryWorkspace, IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
-import { withNullAsUndefined } from 'vs/base/common/types';
-import { IDataTransfer } from 'vs/editor/common/dnd';
-import { extractSelection } from 'vs/platform/opener/common/opener';
+import { DragMouseEvent } from 'vs/base/browser/mouseEvent';
 import { IListDragAndDrop } from 'vs/base/browser/ui/list/list';
 import { ElementsDragAndDropData } from 'vs/base/browser/ui/list/listView';
 import { ITreeDragOverReaction } from 'vs/base/browser/ui/tree/tree';
-import { WebFileSystemAccess } from 'vs/platform/files/browser/webFileSystemAccess';
-import { HTMLFileSystemProvider } from 'vs/platform/files/browser/htmlFileSystemProvider';
-import { DeferredPromise } from 'vs/base/common/async';
+import { coalesce } from 'vs/base/common/arrays';
+import { IDataTransfer } from 'vs/base/common/dataTransfer';
+import { Emitter } from 'vs/base/common/event';
+import { Disposable, DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
+import { stringify } from 'vs/base/common/marshalling';
+import { Mimes } from 'vs/base/common/mime';
+import { FileAccess, Schemas } from 'vs/base/common/network';
+import { isWindows } from 'vs/base/common/platform';
+import { basename, isEqual } from 'vs/base/common/resources';
+import { URI } from 'vs/base/common/uri';
+import { CodeDataTransfers, createDraggedEditorInputFromRawResourcesData, Extensions, extractEditorsDropData, IDragAndDropContributionRegistry, IDraggedResourceEditorInput, IResourceStat } from 'vs/platform/dnd/browser/dnd';
+import { IFileService } from 'vs/platform/files/common/files';
+import { IInstantiationService, ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+import { ILabelService } from 'vs/platform/label/common/label';
+import { extractSelection } from 'vs/platform/opener/common/opener';
 import { Registry } from 'vs/platform/registry/common/platform';
+import { IWindowOpenable } from 'vs/platform/window/common/window';
+import { hasWorkspaceFileExtension, isTemporaryWorkspace, IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
+import { IWorkspaceFolderCreationData, IWorkspacesService } from 'vs/platform/workspaces/common/workspaces';
+import { EditorResourceAccessor, GroupIdentifier, IEditorIdentifier, isEditorIdentifier } from 'vs/workbench/common/editor';
+import { IEditorGroup } from 'vs/workbench/services/editor/common/editorGroupsService';
+import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
+import { IHostService } from 'vs/workbench/services/host/browser/host';
+import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
+import { IWorkspaceEditingService } from 'vs/workbench/services/workspaces/common/workspaceEditing';
 
 //#region Editor / Resources DND
 
@@ -58,124 +50,6 @@ export class DraggedEditorGroupIdentifier {
 export class DraggedTreeItemsIdentifier {
 
 	constructor(readonly identifier: string) { }
-}
-
-export const CodeDataTransfers = {
-	EDITORS: 'CodeEditors',
-	FILES: 'CodeFiles'
-};
-
-export interface IDraggedResourceEditorInput extends IBaseTextResourceEditorInput {
-	resource: URI | undefined;
-
-	/**
-	 * A hint that the source of the dragged editor input
-	 * might not be the application but some external tool.
-	 */
-	isExternal?: boolean;
-
-	/**
-	 * Whether we probe for the dropped editor to be a workspace
-	 * (i.e. code-workspace file or even a folder), allowing to
-	 * open it as workspace instead of opening as editor.
-	 */
-	allowWorkspaceOpen?: boolean;
-}
-
-export async function extractEditorsDropData(accessor: ServicesAccessor, e: DragEvent): Promise<Array<IDraggedResourceEditorInput>> {
-	const editors: IDraggedResourceEditorInput[] = [];
-	if (e.dataTransfer && e.dataTransfer.types.length > 0) {
-
-		// Data Transfer: Code Editors
-		const rawEditorsData = e.dataTransfer.getData(CodeDataTransfers.EDITORS);
-		if (rawEditorsData) {
-			try {
-				editors.push(...parse(rawEditorsData));
-			} catch (error) {
-				// Invalid transfer
-			}
-		}
-
-		// Data Transfer: Resources
-		else {
-			try {
-				const rawResourcesData = e.dataTransfer.getData(DataTransfers.RESOURCES);
-				editors.push(...createDraggedEditorInputFromRawResourcesData(rawResourcesData));
-			} catch (error) {
-				// Invalid transfer
-			}
-		}
-
-		// Check for native file transfer
-		if (e.dataTransfer?.files) {
-			for (let i = 0; i < e.dataTransfer.files.length; i++) {
-				const file = e.dataTransfer.files[i];
-				if (file?.path /* Electron only */) {
-					try {
-						editors.push({ resource: URI.file(file.path), isExternal: true, allowWorkspaceOpen: true });
-					} catch (error) {
-						// Invalid URI
-					}
-				}
-			}
-		}
-
-		// Check for CodeFiles transfer
-		const rawCodeFiles = e.dataTransfer.getData(CodeDataTransfers.FILES);
-		if (rawCodeFiles) {
-			try {
-				const codeFiles: string[] = JSON.parse(rawCodeFiles);
-				for (const codeFile of codeFiles) {
-					editors.push({ resource: URI.file(codeFile), isExternal: true, allowWorkspaceOpen: true });
-				}
-			} catch (error) {
-				// Invalid transfer
-			}
-		}
-
-		// Web: Check for file transfer
-		if (isWeb && containsDragType(e, DataTransfers.FILES)) {
-			const files = e.dataTransfer.items;
-			if (files) {
-				const instantiationService = accessor.get(IInstantiationService);
-				const filesData = await instantiationService.invokeFunction(accessor => extractFilesDropData(accessor, e));
-				for (const fileData of filesData) {
-					editors.push({ resource: fileData.resource, contents: fileData.contents?.toString(), isExternal: true, allowWorkspaceOpen: fileData.isDirectory });
-				}
-			}
-		}
-
-		// Workbench contributions
-		const contributions = Registry.as<IDragAndDropContributionRegistry>(Extensions.DragAndDropContribution).getAll();
-		for (const contribution of contributions) {
-			const data = e.dataTransfer.getData(contribution.dataFormatKey);
-			if (data) {
-				try {
-					editors.push(...contribution.getEditorInputs(data));
-				} catch (error) {
-					// Invalid transfer
-				}
-			}
-		}
-	}
-
-	return editors;
-}
-
-function createDraggedEditorInputFromRawResourcesData(rawResourcesData: string | undefined): IDraggedResourceEditorInput[] {
-	const editors: IDraggedResourceEditorInput[] = [];
-
-	if (rawResourcesData) {
-		const resourcesRaw: string[] = JSON.parse(rawResourcesData);
-		for (const resourceRaw of resourcesRaw) {
-			if (resourceRaw.indexOf(':') > 0) { // mitigate https://github.com/microsoft/vscode/issues/124946
-				const { selection, uri } = extractSelection(URI.parse(resourceRaw));
-				editors.push({ resource: uri, options: { selection } });
-			}
-		}
-	}
-
-	return editors;
 }
 
 export async function extractTreeDropData(dataTransfer: IDataTransfer): Promise<Array<IDraggedResourceEditorInput>> {
@@ -199,121 +73,6 @@ export async function extractTreeDropData(dataTransfer: IDataTransfer): Promise<
 export function convertResourceUrlsToUriList(resourceUrls: string): string {
 	const asJson: URI[] = JSON.parse(resourceUrls);
 	return asJson.map(uri => uri.toString()).join('\n');
-}
-
-interface IFileTransferData {
-	resource: URI;
-	isDirectory?: boolean;
-	contents?: VSBuffer;
-}
-
-async function extractFilesDropData(accessor: ServicesAccessor, event: DragEvent): Promise<IFileTransferData[]> {
-
-	// Try to extract via `FileSystemHandle`
-	if (WebFileSystemAccess.supported(window)) {
-		const items = event.dataTransfer?.items;
-		if (items) {
-			return extractFileTransferData(accessor, items);
-		}
-	}
-
-	// Try to extract via `FileList`
-	const files = event.dataTransfer?.files;
-	if (!files) {
-		return [];
-	}
-
-	return extractFileListData(accessor, files);
-}
-
-async function extractFileTransferData(accessor: ServicesAccessor, items: DataTransferItemList): Promise<IFileTransferData[]> {
-	const fileSystemProvider = accessor.get(IFileService).getProvider(Schemas.file);
-	if (!(fileSystemProvider instanceof HTMLFileSystemProvider)) {
-		return []; // only supported when running in web
-	}
-
-	const results: DeferredPromise<IFileTransferData | undefined>[] = [];
-
-	for (let i = 0; i < items.length; i++) {
-		const file = items[i];
-		if (file) {
-			const result = new DeferredPromise<IFileTransferData | undefined>();
-			results.push(result);
-
-			(async () => {
-				try {
-					const handle = await file.getAsFileSystemHandle();
-					if (!handle) {
-						result.complete(undefined);
-						return;
-					}
-
-					if (WebFileSystemAccess.isFileSystemFileHandle(handle)) {
-						result.complete({
-							resource: await fileSystemProvider.registerFileHandle(handle),
-							isDirectory: false
-						});
-					} else if (WebFileSystemAccess.isFileSystemDirectoryHandle(handle)) {
-						result.complete({
-							resource: await fileSystemProvider.registerDirectoryHandle(handle),
-							isDirectory: true
-						});
-					} else {
-						result.complete(undefined);
-					}
-				} catch (error) {
-					result.complete(undefined);
-				}
-			})();
-		}
-	}
-
-	return coalesce(await Promise.all(results.map(result => result.p)));
-}
-
-export async function extractFileListData(accessor: ServicesAccessor, files: FileList): Promise<IFileTransferData[]> {
-	const dialogService = accessor.get(IDialogService);
-
-	const results: DeferredPromise<IFileTransferData | undefined>[] = [];
-
-	for (let i = 0; i < files.length; i++) {
-		const file = files.item(i);
-		if (file) {
-
-			// Skip for very large files because this operation is unbuffered
-			if (file.size > 100 * ByteSize.MB) {
-				dialogService.show(Severity.Warning, localize('fileTooLarge', "File is too large to open as untitled editor. Please upload it first into the file explorer and then try again."));
-				continue;
-			}
-
-			const result = new DeferredPromise<IFileTransferData | undefined>();
-			results.push(result);
-
-			const reader = new FileReader();
-
-			reader.onerror = () => result.complete(undefined);
-			reader.onabort = () => result.complete(undefined);
-
-			reader.onload = async event => {
-				const name = file.name;
-				const loadResult = withNullAsUndefined(event.target?.result);
-				if (typeof name !== 'string' || typeof loadResult === 'undefined') {
-					result.complete(undefined);
-					return;
-				}
-
-				result.complete({
-					resource: URI.from({ scheme: Schemas.untitled, path: name }),
-					contents: typeof loadResult === 'string' ? VSBuffer.fromString(loadResult) : VSBuffer.wrap(new Uint8Array(loadResult))
-				});
-			};
-
-			// Start reading
-			reader.readAsArrayBuffer(file);
-		}
-	}
-
-	return coalesce(await Promise.all(results.map(result => result.p)));
 }
 
 export interface IResourcesDropHandlerOptions {
@@ -441,11 +200,6 @@ export class ResourcesDropHandler {
 
 		return true;
 	}
-}
-
-interface IResourceStat {
-	resource: URI;
-	isDirectory?: boolean;
 }
 
 export function fillEditorsDragData(accessor: ServicesAccessor, resources: URI[], event: DragMouseEvent | DragEvent): void;
@@ -591,48 +345,6 @@ export function fillEditorsDragData(accessor: ServicesAccessor, resourcesOrEdito
 
 //#endregion
 
-//#region DND contributions
-
-export interface IDragAndDropContributionRegistry {
-	/**
-	 * Registers a drag and drop contribution.
-	 */
-	register(contribution: IDragAndDropContribution): void;
-
-	/**
-	 * Returns all registered drag and drop contributions.
-	 */
-	getAll(): IterableIterator<IDragAndDropContribution>;
-}
-
-export interface IDragAndDropContribution {
-	readonly dataFormatKey: string;
-	getEditorInputs(data: string): IDraggedResourceEditorInput[];
-	setData(resources: IResourceStat[], event: DragMouseEvent | DragEvent): void;
-}
-
-class DragAndDropContributionRegistry implements IDragAndDropContributionRegistry {
-	private readonly _contributions = new Map<string, IDragAndDropContribution>();
-
-	register(contribution: IDragAndDropContribution): void {
-		if (this._contributions.has(contribution.dataFormatKey)) {
-			throw new Error(`A drag and drop contributiont with key '${contribution.dataFormatKey}' was already registered.`);
-		}
-		this._contributions.set(contribution.dataFormatKey, contribution);
-	}
-
-	getAll(): IterableIterator<IDragAndDropContribution> {
-		return this._contributions.values();
-	}
-}
-
-export const Extensions = {
-	DragAndDropContribution: 'workbench.contributions.dragAndDrop'
-};
-
-Registry.add(Extensions.DragAndDropContribution, new DragAndDropContributionRegistry());
-
-//#endregion
 
 //#region DND Utilities
 
@@ -679,26 +391,6 @@ export class LocalSelectionTransfer<T> {
 			this.proto = proto;
 		}
 	}
-}
-
-export function containsDragType(event: DragEvent, ...dragTypesToFind: string[]): boolean {
-	if (!event.dataTransfer) {
-		return false;
-	}
-
-	const dragTypes = event.dataTransfer.types;
-	const lowercaseDragTypes: string[] = [];
-	for (let i = 0; i < dragTypes.length; i++) {
-		lowercaseDragTypes.push(dragTypes[i].toLowerCase()); // somehow the types are lowercase
-	}
-
-	for (const dragType of dragTypesToFind) {
-		if (lowercaseDragTypes.indexOf(dragType.toLowerCase()) >= 0) {
-			return true;
-		}
-	}
-
-	return false;
 }
 
 //#endregion
